Reduction of pub sockets and bug fix at unsuscribing
Instead of having one socket per publisher, only have one socket, and
connect() to multiple publishers in the same socket (zeromq allows this).

Also a bug was fixed that prevented the removal/unregister of the last
websocket per user (was checking position > 0 instead of position >= 0);
which also prevented the unsuscribing of the users.